Cleaned up Signup flow and dead code (#72)
* Removed old homepage

* Remove other death row items

* Removed MedicalCode

* removed death row item

* update the patients api

* Fixed the update and create items

* Updated the responses from server

* Clean up signup flow

* Change the Devices View Controller

* Cleaned up the notifications view controller

* Removed TOS and PP reduntent controllers

* Adjusted the buttons

* Updated the flow for signup

* Cleaned up Webserver

+++ Allie/UIElements/Custom Views/ProfileChartView.swift
@@ -49,7 +49,7 @@ class ProfileChartView: LineChartView {
 		xAxis.labelFont = .boldSystemFont(ofSize: 13.0)
 	}
 
-	func setup(with healthData: [StatModel], quantityType: HealthKitQuantityType, intervalType: HealthStatsDateIntervalType, goal: Double) {
+	func setup(with healthData: [StatModel], quantityType: HealthKitQuantityType, intervalType: HealthStatsDateIntervalType, goal: Int) {
 		guard let statModel = healthData.first, let firstDate = statModel.dataPoints.first?.date, let lastDate = statModel.dataPoints.last?.date else { data = nil; return }
 
 		rightAxis.enabled = false
@@ -153,8 +153,8 @@ class ProfileChartView: LineChartView {
 		}
 	}
 
-	func setUpBaseline(min: Double, max: Double, dataSets: [LineChartDataSet], goal: Double) {
-		baseLine.limit = goal
+	func setUpBaseline(min: Double, max: Double, dataSets: [LineChartDataSet], goal: Int) {
+		baseLine.limit = Double(goal)
 		leftAxis.axisMinimum = baseLine.limit < min ? baseLine.limit - 10.0 : min
 		leftAxis.axisMaximum = baseLine.limit > max ? baseLine.limit + 10.0 : max
 		baseLine.lineWidth = 1
